requirements.txt
aiohttp==3.5.4
python-socketio==4.0.0
requests==2.21.0
